+++ app/src/main/res/layout/fragment_group.xml
@@ -26,8 +26,6 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content" />
 
-
-
         <android.support.v7.widget.RecyclerView
             android:id="@+id/group_recycler_view"
             android:scrollbars="vertical"
